backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
anthropic
stripe
python-dotenv
google-auth
google-generativeai
requests
eval-type-backport
bcrypt
python-jose[cryptography]
